# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/data/idrange": {
            "get": {
                "description": "get a range of ids in a given time range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data"
                ],
                "summary": "id range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "start date",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "end date",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Dataset"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorMsg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Dataset": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Actual data\nexample: { \"key\": \"dataset1\" }",
                    "type": "integer"
                },
                "id": {
                    "description": "The ID of a dataaset\nexample: 130670",
                    "type": "string"
                }
            }
        },
        "models.ErrorMsg": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}